Do not compress the UVFITS and FITS files
Compressed (with gzip) UVFITS and FITS can't be read by IMAGER
directly; they need to be uncompressed first, which takes a lot of
time because the files are large. In addition, they can't be read in
memory-mapping mode in Python, which means that the entire cube needs
to be loaded in RAM for each operation. It is better not to compress them.