Store the noise computed from the clean image/cube
The noise computed from primary-corrected images overestimates the
noise level at the center of the field-of-view. In addition it cannot
be compared to the thermal noise to determine if the image/cube is
limited by the dynamical range. On the other hand, the noise computed
from the uncorrected image/cubes is a good estimate of the noise at
the center of primary-beam corrected images. Store this value instead.