Compute the continuum in the image plane
The continuum is now computed from the cleaned data cube. By default,
it is computed using the SCM algorithm, which is based on the
per-pixel spectrum. Each spectrum is sigma-clipped until convergence
of the noise. The continuum image is subtracted from cleaned data cube
to produce the line data cube.